Fixed some errors with MEM_TEST=true

+++ include/galsim/mmgr.h
@@ -10,120 +10,123 @@
 // [NOTE: Best when viewed with 8-character tabs]
 // ---------------------------------------------------------------------------------------------------------------------------------
 
-#ifndef	_H_MMGR
-#define	_H_MMGR
+#ifndef _H_MMGR
+#define _H_MMGR
 
 // ---------------------------------------------------------------------------------------------------------------------------------
 // For systems that don't have the __FUNCTION__ variable, we can just define it here
 // ---------------------------------------------------------------------------------------------------------------------------------
 
-//#define	__FUNCTION__ __func__
+//#define __FUNCTION__ __func__
 
 // ---------------------------------------------------------------------------------------------------------------------------------
 // Types
 // ---------------------------------------------------------------------------------------------------------------------------------
 
-typedef	struct tag_au
+typedef struct tag_au
 {
-	size_t		actualSize;
-	size_t		reportedSize;
-	void		*actualAddress;
-	void		*reportedAddress;
-	char		sourceFile[40];
-	char		sourceFunc[40];
-	unsigned int	sourceLine;
-	unsigned int	allocationType;
-	bool		breakOnDealloc;
-	bool		breakOnRealloc;
-	unsigned int	allocationNumber;
-	struct tag_au	*next;
-	struct tag_au	*prev;
+    size_t actualSize;
+    size_t reportedSize;
+    void  *actualAddress;
+    void  *reportedAddress;
+    char  sourceFile[40];
+    char  sourceFunc[40];
+    size_t sourceLine;
+    size_t allocationType;
+    bool  breakOnDealloc;
+    bool  breakOnRealloc;
+    size_t allocationNumber;
+    struct tag_au *next;
+    struct tag_au *prev;
 } sAllocUnit;
 
-typedef	struct
+typedef struct
 {
-	unsigned int	totalReportedMemory;
-	unsigned int	totalActualMemory;
-	unsigned int	peakReportedMemory;
-	unsigned int	peakActualMemory;
-	unsigned int	accumulatedReportedMemory;
-	unsigned int	accumulatedActualMemory;
-	unsigned int	accumulatedAllocUnitCount;
-	unsigned int	totalAllocUnitCount;
-	unsigned int	peakAllocUnitCount;
+    size_t totalReportedMemory;
+    size_t totalActualMemory;
+    size_t peakReportedMemory;
+    size_t peakActualMemory;
+    size_t accumulatedReportedMemory;
+    size_t accumulatedActualMemory;
+    size_t accumulatedAllocUnitCount;
+    size_t totalAllocUnitCount;
+    size_t peakAllocUnitCount;
 } sMStats;
 
 // ---------------------------------------------------------------------------------------------------------------------------------
 // External constants
 // ---------------------------------------------------------------------------------------------------------------------------------
 
-extern	const	unsigned int	m_alloc_unknown;
-extern	const	unsigned int	m_alloc_new;
-extern	const	unsigned int	m_alloc_new_array;
-extern	const	unsigned int	m_alloc_malloc;
-extern	const	unsigned int	m_alloc_calloc;
-extern	const	unsigned int	m_alloc_realloc;
-extern	const	unsigned int	m_alloc_delete;
-extern	const	unsigned int	m_alloc_delete_array;
-extern	const	unsigned int	m_alloc_free;
+extern const size_t m_alloc_unknown;
+extern const size_t m_alloc_new;
+extern const size_t m_alloc_new_array;
+extern const size_t m_alloc_malloc;
+extern const size_t m_alloc_calloc;
+extern const size_t m_alloc_realloc;
+extern const size_t m_alloc_delete;
+extern const size_t m_alloc_delete_array;
+extern const size_t m_alloc_free;
 
 // ---------------------------------------------------------------------------------------------------------------------------------
 // Used by the macros
 // ---------------------------------------------------------------------------------------------------------------------------------
 
-void		m_setOwner(const char *file, const unsigned int line, const char *func);
+void m_setOwner(const char *file, const size_t line, const char *func);
 
 // ---------------------------------------------------------------------------------------------------------------------------------
 // Allocation breakpoints
 // ---------------------------------------------------------------------------------------------------------------------------------
 
-bool		&m_breakOnRealloc(void *reportedAddress);
-bool		&m_breakOnDealloc(void *reportedAddress);
+bool &m_breakOnRealloc(void *reportedAddress);
+bool &m_breakOnDealloc(void *reportedAddress);
 
 // ---------------------------------------------------------------------------------------------------------------------------------
 // The meat of the memory tracking software
 // ---------------------------------------------------------------------------------------------------------------------------------
 
-void		*m_allocator(const char *sourceFile, const unsigned int sourceLine, const char *sourceFunc,
-			     const unsigned int allocationType, const size_t reportedSize);
-void		*m_reallocator(const char *sourceFile, const unsigned int sourceLine, const char *sourceFunc,
-			       const unsigned int reallocationType, const size_t reportedSize, void *reportedAddress);
-void		m_deallocator(const char *sourceFile, const unsigned int sourceLine, const char *sourceFunc,
-			      const unsigned int deallocationType, const void *reportedAddress);
+void *m_allocator(
+    const char *sourceFile, const size_t sourceLine, const char *sourceFunc,
+    const size_t allocationType, const size_t reportedSize);
+void *m_reallocator(
+    const char *sourceFile, const size_t sourceLine, const char *sourceFunc,
+    const size_t reallocationType, const size_t reportedSize, void *reportedAddress);
+void  m_deallocator(
+    const char *sourceFile, const size_t sourceLine, const char *sourceFunc,
+    const size_t deallocationType, const void *reportedAddress);
 
 // ---------------------------------------------------------------------------------------------------------------------------------
 // Utilitarian functions
 // ---------------------------------------------------------------------------------------------------------------------------------
 
-bool		m_validateAddress(const void *reportedAddress);
-bool		m_validateAllocUnit(const sAllocUnit *allocUnit);
-bool		m_validateAllAllocUnits();
+bool m_validateAddress(const void *reportedAddress);
+bool m_validateAllocUnit(const sAllocUnit *allocUnit);
+bool m_validateAllAllocUnits();
 
 // ---------------------------------------------------------------------------------------------------------------------------------
 // Unused RAM calculations
 // ---------------------------------------------------------------------------------------------------------------------------------
 
-unsigned int	m_calcUnused(const sAllocUnit *allocUnit);
-unsigned int	m_calcAllUnused();
+size_t m_calcUnused(const sAllocUnit *allocUnit);
+size_t m_calcAllUnused();
 
 // ---------------------------------------------------------------------------------------------------------------------------------
 // Logging and reporting
 // ---------------------------------------------------------------------------------------------------------------------------------
 
-void		m_dumpAllocUnit(const sAllocUnit *allocUnit, const char *prefix = "");
-void		m_dumpMemoryReport(const char *filename = "memreport.log", const bool overwrite = true);
-sMStats		m_getMemoryStatistics();
+void m_dumpAllocUnit(const sAllocUnit *allocUnit, const char *prefix = "");
+void m_dumpMemoryReport(const char *filename = "memreport.log", const bool overwrite = true);
+sMStats m_getMemoryStatistics();
 
 // ---------------------------------------------------------------------------------------------------------------------------------
 // Variations of global operators new & delete
 // ---------------------------------------------------------------------------------------------------------------------------------
 
-void	*operator new(size_t reportedSize) throw(std::bad_alloc);
-void	*operator new[](size_t reportedSize) throw(std::bad_alloc);
-void	*operator new(size_t reportedSize, const char *sourceFile, int sourceLine) throw(std::bad_alloc);
-void	*operator new[](size_t reportedSize, const char *sourceFile, int sourceLine) throw(std::bad_alloc);
-void	operator delete(void *reportedAddress) throw();
-void	operator delete[](void *reportedAddress) throw();
+void *operator new(size_t reportedSize) throw(std::bad_alloc);
+void *operator new[](size_t reportedSize) throw(std::bad_alloc);
+void *operator new(size_t reportedSize, const char *sourceFile, int sourceLine) throw(std::bad_alloc);
+void *operator new[](size_t reportedSize, const char *sourceFile, int sourceLine) throw(std::bad_alloc);
+void operator delete(void *reportedAddress) throw();
+void operator delete[](void *reportedAddress) throw();
 
 #endif // _H_MMGR
 
@@ -132,12 +135,12 @@ void	operator delete[](void *reportedAddress) throw();
 // ---------------------------------------------------------------------------------------------------------------------------------
 
 #include "nommgr.h"
-#define	new		(m_setOwner  (__FILE__,__LINE__,__FUNCTION__),false) ? NULL : new
-#define	delete		(m_setOwner  (__FILE__,__LINE__,__FUNCTION__),false) ? m_setOwner("",0,"") : delete
-#define	malloc(sz)	m_allocator  (__FILE__,__LINE__,__FUNCTION__,m_alloc_malloc,sz)
-#define	calloc(sz)	m_allocator  (__FILE__,__LINE__,__FUNCTION__,m_alloc_calloc,sz)
-#define	realloc(ptr,sz)	m_reallocator(__FILE__,__LINE__,__FUNCTION__,m_alloc_realloc,sz,ptr)
-#define	free(ptr)	m_deallocator(__FILE__,__LINE__,__FUNCTION__,m_alloc_free,ptr)
+#define new (m_setOwner (__FILE__,__LINE__,__FUNCTION__),false) ? NULL : new
+#define delete (m_setOwner (__FILE__,__LINE__,__FUNCTION__),false) ? m_setOwner("",0,"") : delete
+#define malloc(sz) m_allocator (__FILE__,__LINE__,__FUNCTION__,m_alloc_malloc,sz)
+#define calloc(sz) m_allocator (__FILE__,__LINE__,__FUNCTION__,m_alloc_calloc,sz)
+#define realloc(ptr,sz) m_reallocator(__FILE__,__LINE__,__FUNCTION__,m_alloc_realloc,sz,ptr)
+#define free(ptr) m_deallocator(__FILE__,__LINE__,__FUNCTION__,m_alloc_free,ptr)
 
 // ---------------------------------------------------------------------------------------------------------------------------------
 // mmgr.h - End of file
